# perf: lazy load all pages