--- a/src/components/Layout/Layout.tsx
+++ b/src/components/Layout/Layout.tsx
@@ -1,6 +1,6 @@
 import type { ContainerProps } from '@chakra-ui/react'
 import { Container, Flex } from '@chakra-ui/react'
-import React, { Suspense } from 'react'
+import React from 'react'
 
 import { Header } from './Header/Header'
 import { SideNav } from './Header/SideNav'
@@ -26,7 +26,7 @@ export const Layout: React.FC<ContainerProps> = ({ children, ...rest }) => {
         {...rest}
       >
         <Header />
-        <Suspense>{children}</Suspense>
+        {children}
       </Container>
     </Flex>
   )