… src/context/ModalProvider/types.ts