## Description

Adds lazy loading to all pages and modals, with spinner so app doesn't look broken during load (as is currently with trade page).

Improves script eval marginally for faster time to interactive which is something that desperately needs improving for mobile devices.

## Pull Request Type

- [ ] :bug: Bug fix (Non-breaking Change: Fixes an issue)
- [x] :hammer_and_wrench: Chore (Non-breaking Change: Doc updates, pkg upgrades, typos, etc..)
- [ ] :nail_care: New Feature (Breaking/Non-breaking Change)

## Issue (if applicable)

NA

## Risk
> High Risk PRs Require 2 approvals

Low/moderate risk of pages or modals not displaying at all, or the spinner somehow appearing in an unexpected way.

> What protocols, transaction types or contract interactions might be affected by this PR?

## Testing

Check all pages and modals are loading and initially displaying as intended. Also check wallet connect dapps drop-down is loading. 

There will be a spinner on the first load which is expected - the idea here is we're moving the load time to when a page is first viewed rather than everything on app boot which improves load time for most users.

### Engineering

<!-----------------------------------------------------------------------------
Include sufficient information here for an engineer to test your PR. This may include how to test locally, in a built environment, changes to infrastructure etc.
------------------------------------------------------------------------------>

### Operations

<!-----------------------------------------------------------------------------
If your changes have a user-facing impact, describe how a non-technical QA team can functionally test your changes in a preview environment.

If they are not user-facing please describe how to test for any regressions that may occur.
------------------------------------------------------------------------------>

## Screenshots (if applicable)

Spinner during lazy load
https://github.com/shapeshift/web/assets/125113430/bbea3f98-d91c-418d-a5cc-4cd1ff5c4cdc

Before this PR - 1.51s to parse, DCL at ~2.4s
![image](https://github.com/shapeshift/web/assets/125113430/5c640346-5fb6-4346-b60d-ce573ba3530a)

This PR - 1.17s to parse, DCL at ~1.9s
![image](https://github.com/shapeshift/web/assets/125113430/375664fe-94ea-42bf-a797-e4d01f5221b1)