@woodenfurniture 2024-02-08
Opted to avoid makeSuspenseful util in the other 2 cases for now, as the spinner has been styled for full width which results in wierd styling:
![image](https://github.com/shapeshift/web/assets/125113430/7419956a-638f-41b9-886f-e44ebb9c72c1)